+++ res/xml/development_settings.xml
@@ -492,10 +492,6 @@
             android:summary="@string/hwui_force_dark_summary" />
 
         <SwitchPreference
-            android:key="enable_blurs_on_windows"
-            android:title="@string/window_blurs"/>
-
-        <SwitchPreference
             android:key="force_msaa"
             android:title="@string/force_msaa"
             android:summary="@string/force_msaa_summary" />
